MXS-4463 Add base_read_buffer_size config paramater
This is at least initially intended for profiling and
memory usage debugging.

Currently DCB initially allocates a 64kB buffer for reading.
That means there is quite an overhead, e.g. for a simple
returned OK packet. This may contribute to the increased
memory usage (to the extent that it causes OOMs) of 22.08.

With this change, the initial size of the buffer can be
adjusted via the configuration and also at runtime. E.g.

    base_read_buffer_size=32kB

Futher, if 'base_read_buffer_size' is set to 0, then the
initial size of the buffer is set to what ioctl(FIONREAD)
reports for the socket. That's how it used to be.

Changing the initial buffer size can be made like

    maxctrl alter maxscale base_read_buffer_size=0