Add debug=fast-read-size
The debug option allows the size of the first "fast read" to be
configured. If set to a non-zero value and if combined with
base_read_buffer_size=0, the first read is done with the fast read size
and subsequent reads use ioctl calls to peek into the socket buffer.